Tests: fixes tests
