+++ public/app/features/variables/state/reducers.test.ts
@@ -1,10 +1,28 @@
 import { reducerTester } from '../../../../test/core/redux/reducerTester';
 import { cleanUpDashboard } from 'app/features/dashboard/state/reducers';
-import { ConstantVariableModel, QueryVariableModel, VariableHide } from '../../templating/variable';
+import { QueryVariableModel, VariableHide, VariableType } from '../../templating/variable';
 import { VariableAdapter, variableAdapters } from '../adapters';
 import { createAction } from '@reduxjs/toolkit';
 import { variablesReducer, VariablesState } from './variablesReducer';
 import { toVariablePayload, VariablePayload } from './types';
+
+const variableAdapter: VariableAdapter<QueryVariableModel> = {
+  id: 'mock',
+  name: 'Mock label',
+  description: 'Mock description',
+  dependsOn: jest.fn(),
+  updateOptions: jest.fn(),
+  initialState: {} as QueryVariableModel,
+  reducer: jest.fn().mockReturnValue({}),
+  getValueForUrl: jest.fn(),
+  getSaveModel: jest.fn(),
+  picker: null as any,
+  editor: null as any,
+  setValue: jest.fn(),
+  setValueFromUrl: jest.fn(),
+};
+
+variableAdapters.setInit(() => [{ ...variableAdapter }]);
 
 describe('variablesReducer', () => {
   describe('when cleanUpDashboard is dispatched', () => {
@@ -91,31 +109,16 @@ describe('variablesReducer', () => {
           skipUrlSync: false,
         },
       };
-      const variableAdapter: VariableAdapter<QueryVariableModel> = {
-        id: 'mock',
-        name: 'Mock label',
-        description: 'Mock description',
-        dependsOn: jest.fn(),
-        updateOptions: jest.fn(),
-        initialState: {} as QueryVariableModel,
-        reducer: jest.fn().mockReturnValue(initialState),
-        getValueForUrl: jest.fn(),
-        getSaveModel: jest.fn(),
-        picker: null as any,
-        editor: null as any,
-        setValue: jest.fn(),
-        setValueFromUrl: jest.fn(),
-      };
-      variableAdapters.setInit(() => [{ ...variableAdapter }]);
+      variableAdapters.get('mock').reducer = jest.fn().mockReturnValue(initialState);
       const mockAction = createAction<VariablePayload>('mockAction');
       reducerTester<VariablesState>()
         .givenReducer(variablesReducer, initialState)
-        .whenActionIsDispatched(mockAction(toVariablePayload({ type: 'query', uuid: '0' })))
+        .whenActionIsDispatched(mockAction(toVariablePayload({ type: ('mock' as unknown) as VariableType, uuid: '0' })))
         .thenStateShouldEqual(initialState);
-      expect(variableAdapter.reducer).toHaveBeenCalledTimes(1);
-      expect(variableAdapter.reducer).toHaveBeenCalledWith(
+      expect(variableAdapters.get('mock').reducer).toHaveBeenCalledTimes(1);
+      expect(variableAdapters.get('mock').reducer).toHaveBeenCalledWith(
         initialState,
-        mockAction(toVariablePayload({ type: 'query', uuid: '0' }))
+        mockAction(toVariablePayload({ type: ('mock' as unknown) as VariableType, uuid: '0' }))
       );
     });
   });
@@ -133,28 +136,13 @@ describe('variablesReducer', () => {
           skipUrlSync: false,
         },
       };
-      const variableAdapter: VariableAdapter<QueryVariableModel> = {
-        id: 'mock',
-        name: 'Mock label',
-        description: 'Mock description',
-        dependsOn: jest.fn(),
-        updateOptions: jest.fn(),
-        initialState: {} as QueryVariableModel,
-        reducer: jest.fn().mockReturnValue(initialState),
-        getValueForUrl: jest.fn(),
-        getSaveModel: jest.fn(),
-        picker: null as any,
-        editor: null as any,
-        setValue: jest.fn(),
-        setValueFromUrl: jest.fn(),
-      };
-      variableAdapters.setInit(() => [{ ...variableAdapter }]);
+      variableAdapters.get('mock').reducer = jest.fn().mockReturnValue(initialState);
       const mockAction = createAction<VariablePayload>('mockAction');
       reducerTester<VariablesState>()
         .givenReducer(variablesReducer, initialState)
         .whenActionIsDispatched(mockAction(toVariablePayload({ type: 'adhoc', uuid: '0' })))
         .thenStateShouldEqual(initialState);
-      expect(variableAdapter.reducer).toHaveBeenCalledTimes(0);
+      expect(variableAdapters.get('mock').reducer).toHaveBeenCalledTimes(0);
     });
   });
 
@@ -171,28 +159,13 @@ describe('variablesReducer', () => {
           skipUrlSync: false,
         },
       };
-      const variableAdapter: VariableAdapter<ConstantVariableModel> = {
-        id: 'mock',
-        name: 'Mock label',
-        description: 'Mock description',
-        dependsOn: jest.fn(),
-        updateOptions: jest.fn(),
-        initialState: {} as ConstantVariableModel,
-        reducer: jest.fn().mockReturnValue(initialState),
-        getValueForUrl: jest.fn(),
-        getSaveModel: jest.fn(),
-        picker: null as any,
-        editor: null as any,
-        setValue: jest.fn(),
-        setValueFromUrl: jest.fn(),
-      };
-      variableAdapters.setInit(() => [{ ...variableAdapter }]);
+      variableAdapters.get('mock').reducer = jest.fn().mockReturnValue(initialState);
       const mockAction = createAction<string>('mockAction');
       reducerTester<VariablesState>()
         .givenReducer(variablesReducer, initialState)
         .whenActionIsDispatched(mockAction('mocked'))
         .thenStateShouldEqual(initialState);
-      expect(variableAdapter.reducer).toHaveBeenCalledTimes(0);
+      expect(variableAdapters.get('mock').reducer).toHaveBeenCalledTimes(0);
     });
   });
 });
